// Frontend/src/generated/model/productsIdBody.ts
/**
 * SportsHub
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface ProductsIdBody { 
    name?: any;
    description?: any;
    price?: any;
    category?: any;
    brand?: any;
    imageUrl?: any;
}